Change: Applied patch attached to issue #67 by myk.taylor (myk0002) which solves several compiler warnings and one potential bug. (Thanks myk.taylor!)  - Closes issue #67
git-svn-id: http://better-oblivion-sorting-software.googlecode.com/svn@1654 bfee223e-1fc4-11de-b650-e715bd6d7cf1

+++ code/boss-common/trunk/Support/Helpers.cpp
@@ -126,7 +126,7 @@ namespace boss {
 	}
 
 	string Tidy(string filename) {						//Changes uppercase to lowercase and removes trailing spaces to do what Windows filesystem does to filenames.	
-		int endpos = filename.find_last_not_of(" \t");
+		size_t endpos = filename.find_last_not_of(" \t");
 	
 		if (filename.npos == endpos) return (""); 			//sanity check for empty string
 		else {
